suggested performance improvement
Our production Bareos database is pretty sizeable (~5 TB) and the exporter did not finish in any acceptable timeframe (at least 30+ minutes for us). The biggest issue was the PoolInfo query. With this little improvement the query runs in less than a second and produces the same results. Now the whole scrape takes 2-3 minutes, which is far more manageable.